Refactor project structure again
