+++ SSC-C/src/main/java/cz/muni/fi/sscc/Main.java
@@ -4,10 +4,10 @@ import antlr.SSCLexer;
 import antlr.SSCParser;
 import cz.muni.fi.sscc.exceptions.AntlrException;
 import cz.muni.fi.sscc.exceptions.SSCSyntaxException;
-import cz.muni.fi.sscc.mine.ConvertorVisitor;
-import cz.muni.fi.sscc.mine.PostfixExpressionConvertorVisitor;
-import cz.muni.fi.sscc.mine.SSVisitor;
-import cz.muni.fi.sscc.mine.data.SuperStructRepre;
+import cz.muni.fi.sscc.visitors.ConvertorVisitor;
+import cz.muni.fi.sscc.visitors.PostfixExpressionConvertorVisitor;
+import cz.muni.fi.sscc.visitors.SSVisitor;
+import cz.muni.fi.sscc.data.SuperStructRepre;
 import org.antlr.v4.runtime.CharStreams;
 import org.antlr.v4.runtime.CommonTokenStream;
 import org.antlr.v4.runtime.tree.ParseTree;
