+++ SSC-C/src/main/java/cz/muni/fi/sscc/data/SSMember.java
@@ -1,4 +1,4 @@
-package cz.muni.fi.sscc.mine.data;
+package cz.muni.fi.sscc.data;
 
 import cz.muni.fi.sscc.util.Either;
