+++ SSC-C/src/main/java/cz/muni/fi/sscc/data/SuperStructRepre.java
@@ -1,4 +1,4 @@
-package cz.muni.fi.sscc.mine.data;
+package cz.muni.fi.sscc.data;
 
 import java.util.List;
 import java.util.Objects;
