+++ SSC-C/src/main/java/cz/muni/fi/sscc/visitors/PostfixExpressionConvertorVisitor.java
@@ -1,12 +1,12 @@
-package cz.muni.fi.sscc.mine;
+package cz.muni.fi.sscc.visitors;
 
 import cz.muni.fi.sscc.Main;
 import cz.muni.fi.sscc.exceptions.SSCSyntaxException;
 import antlr.SSCParser;
-import cz.muni.fi.sscc.mine.data.FunctionDefinition;
-import cz.muni.fi.sscc.mine.data.SSMember;
-import cz.muni.fi.sscc.mine.data.SuperStructRepre;
-import cz.muni.fi.sscc.mine.data.SuperstructVariable;
+import cz.muni.fi.sscc.data.FunctionDefinition;
+import cz.muni.fi.sscc.data.SSMember;
+import cz.muni.fi.sscc.data.SuperStructRepre;
+import cz.muni.fi.sscc.data.SuperstructVariable;
 import org.antlr.v4.runtime.CommonTokenStream;
 import org.antlr.v4.runtime.ParserRuleContext;
 import org.antlr.v4.runtime.tree.TerminalNode;
